+++ Paradigm/Changes.md
@@ -18,6 +18,8 @@
   - Added `.relativize(...)` and `.relativeToAbsolute(...)` methods that perform **Bounds**-based coordinate conversions
 - **Color**
   - Added `.visible` and `.invisible`
+- **DimensionalFactory**
+  - Added a new `.twice(D)` constructor
 - **Distance** (object)
   - Added `.ofKilometers(Double)`
 - **NumericVectorLike**